make getIntl a helper method
