--- src/helpers/i18nToUpperCase.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export function i18nToUpperCase(txt) {
-  return (<span>
-    {txt.toUpperCase()}
-  </span>);
-}
